Черновой коммит с частично работающим кодом

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserServiceImpl.java
@@ -9,6 +9,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 
 @Service
@@ -16,13 +17,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class UserServiceImpl implements IUserService {
 
     private final IUserStorage userStorage;
-    private static final AtomicInteger idSequence = new AtomicInteger(1);
+    private static final AtomicLong idSequence = new AtomicLong(1);
 
     /* При создании значение поля User.id обязано быть пустым: управлять идентификаторами может исключительно сервер,
     а значит, клиент обязан прислать данные без идентификатора. Поэтому бездумно вешать валидацию на null нельзя. */
     @Override
     public User add(User user) {
-        int userId = idSequence.getAndIncrement();
+        Long userId = idSequence.getAndIncrement();
         user.setId(userId);
         userStorage.add(user);
         return user;
@@ -38,7 +39,7 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public User get(Integer id) {
+    public User get(Long id) {
         return Optional.ofNullable(userStorage.get(id)).orElseThrow(
                 () -> new NotFoundException("Пользователь с ID " + id + " не найден")
         );
@@ -51,23 +52,23 @@ public class UserServiceImpl implements IUserService {
 
     // добавление в друзья
     @Override
-    public void setFriendship(Integer userId, Integer friendUserId) {
+    public void setFriendship(Long userId, Long friendUserId) {
         User user = get(userId), friendUser = get(friendUserId);
         userStorage.setFriendship(user.getId(), friendUser.getId());
     }
 
     // удаление из друзей
     @Override
-    public void unsetFriendship(Integer userId, Integer friendUserId) {
+    public void unsetFriendship(Long userId, Long friendUserId) {
         User user = get(userId), friendUser = get(friendUserId);
         userStorage.unsetFriendship(user.getId(), friendUser.getId());
     }
 
     // вывод списка общих друзей
     @Override
-    public List<User> getMutualFriends(Integer user1Id, Integer user2Id) {
+    public List<User> getMutualFriends(Long user1Id, Long user2Id) {
         User user1 = get(user1Id), user2 = get(user2Id);
-        Set<Integer> user1FriendsIds = userStorage.getUserFriendsIds(user1.getId()),
+        Set<Long> user1FriendsIds = userStorage.getUserFriendsIds(user1.getId()),
                 user2FriendsIds = userStorage.getUserFriendsIds(user2.getId());
         if (!user1FriendsIds.isEmpty() && !user2FriendsIds.isEmpty())
             user1FriendsIds.retainAll(user2FriendsIds);
@@ -77,7 +78,7 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public List<User> getUserFriends(Integer userId) {
+    public List<User> getUserFriends(Long userId) {
         User user = get(userId);
         return userStorage.getUserFriendsIds(user.getId()).stream()
                 .map(userStorage::get)
